Finish support for 2.9 Bind. Change Bonus and Other points from short to double (needed for NADAC) Fixed a bug in compile.py (don't store full path name - dev only issue)

+++ AgilityBook/src/ARB/ARBDogRun.cpp
@@ -548,7 +548,7 @@ double ARBDogRun::GetTitlePoints(
 		else
 			*outPlacement = 0.0;
 	}
-	short bonusPts = inScoring->HasBonusPts() ? m_Scoring.GetBonusPts() : 0;
+	double bonusPts = inScoring->HasBonusPts() ? m_Scoring.GetBonusPts() : 0.0;
 	switch (m_Scoring.GetType())
 	{
 	default:
